modification interface des bouttons, interface Rolling et garage terminées, ajout d'un background sur plusieurs interfaces, ajout des boutons utiles, doit corriger problème owner pour invités

+++ app/src/main/java/com/sidali/projet/DeviceData.kt
@@ -6,6 +6,6 @@ data class DeviceData(
 val id:String,
     val type:String,
     val availableCommands: ArrayList<String>,
-    val opening : Int?,
+    val opening : Float?,
     val power : Int?
 )
